keep document's members order when doing an insert
By using a bson_doc instead of a perl hash, we keep the items in
order. This is not a big deal for most users because the order
is lost anyway whe you retrieve the document (Mango gives you a
hash) but it can be useful in some environments.
Thanks to Stefan Althöfer for finding this.